pyproject.toml
[tool.poetry]
name = "fai-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Unviray <unviray@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytgcalls = {extras = ["pyrogram"], version = "^2.1.0"}
TgCrypto = "^1.2.2"
Telethon = "^1.23.0"
python-slugify = "^5.0.2"
python-dotenv = "^0.19.0"
rich = "^10.11.0"
textual = "^0.1.12"
prompt-toolkit = "^3.0.20"
nest-asyncio = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
icecream = "^2.1.1"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
